moved options one level up; update import statement
this is to ensure that all enumerators are visible to src.pybella.utils.user_data
